+++ lib/td_df_web/router.ex
@@ -32,7 +32,7 @@ defmodule TdDfWeb.Router do
     pipe_through [:api, :api_secure, :api_authorized]
 
     resources "/templates", TemplateController, except: [:new, :edit]
-    get "/templates/load/:id", TemplateController, :load_and_show
+    #get "/templates/load/:id", TemplateController, :load_and_show
   end
 
   def swagger_info do